enums1

+++ exercises/enums/enums2.rs
@@ -3,13 +3,16 @@
 // Execute `rustlings hint enums2` or use the `hint` watch subcommand for a
 // hint.
 
-
 #[derive(Debug)]
 enum Message {
-    Move{x: u8,y: u8},
+    // TODO: define the different variants used below
+    Move {
+        x: i32,
+        y: i32,
+    },
     Echo(String),
-    ChangeColor(i32, i32, i32),
-    Quit
+    ChangeColor(u32, u32, u32),
+    Quit,
 }
 
 impl Message {
